// src/KeyLens.App/src/api/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type Config = {
    audience: string;
    scope: string;
};

export type CredentialKind = 'Secret' | 'Certificate' | 'Password' | 'Key';

export type CredentialRecord = {
    provider: string;
    container: string;
    containerId: string;
    credentialId: string;
    kind: CredentialKind;
    name: string | null;
    notBefore: string | null;
    expiresOn: string | null;
    enabled: boolean;
    credentialUri?: string | null;
    metadata?: unknown;
};

export type GetConfigData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/api/v1/config';
};

export type GetConfigResponses = {
    /**
     * OK
     */
    200: Config;
};

export type GetConfigResponse = GetConfigResponses[keyof GetConfigResponses];

export type GetCredentialsData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/api/v1/credentials';
};

export type GetCredentialsResponses = {
    /**
     * OK
     */
    200: Array<CredentialRecord>;
};

export type GetCredentialsResponse = GetCredentialsResponses[keyof GetCredentialsResponses];

export type ClientOptions = {
    baseURL: 'http://localhost:5011/' | (string & {});
};